requirements.txt
beautifulsoup4==4.10.0
Jinja2==3.0.3
Pillow==9.0.1
pyunsplash==1.0.0rc2
requests==2.27.1